# server/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Valkyrie REST API Specs. This service uses sessions for authentication",
        "title": "Valkyrie API",
        "contact": {},
        "license": {
            "name": "Apache 2.0"
        },
        "version": "1.0"
    },
    "host": "localhost:\u003cPORT\u003e",
    "basePath": "/api",
    "paths": {
...
        "/messages/{channelId}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Messages"
                ],
                "summary": "Get Channel Messages",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Channel ID",
                        "name": "channelId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Cursor Pagination using the createdAt field",
                        "name": "cursor",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/Message"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Messages"
                ],
                "summary": "Create Messages",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Channel ID",
                        "name": "channelId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Create Message",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/MessageRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/ErrorsResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    }
                }
            }
        },
        "/messages/{messageId}": {
            "put": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Messages"
                ],
                "summary": "Edit Messages",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Message ID",
                        "name": "messageId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Edit Message",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/MessageRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/ErrorsResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Messages"
                ],
                "summary": "Delete Messages",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Message ID",
                        "name": "messageId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/SuccessResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "Attachment": {
            "type": "object",
            "properties": {
                "filename": {
                    "type": "string"
                },
                "filetype": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                }
            }
        },
...
        "ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "$ref": "#/definitions/HttpError"
                }
            }
        },
        "ErrorsResponse": {
            "type": "object",
            "properties": {
                "errors": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/FieldError"
                    }
                }
            }
        },
        "FieldError": {
            "type": "object",
            "properties": {
                "field": {
                    "description": "The property containing the error",
                    "type": "string"
                },
                "message": {
                    "description": "The specific error message",
                    "type": "string"
                }
            }
        },
...
        "HttpError": {
            "type": "object",
            "properties": {
                "message": {
                    "description": "The specific error message",
                    "type": "string"
                },
                "type": {
                    "description": "The Http Response as a string",
                    "type": "string"
                }
            }
        },
...
        "Member": {
            "type": "object",
            "properties": {
                "color": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "image": {
                    "type": "string"
                },
                "isFriend": {
                    "type": "boolean"
                },
                "isOnline": {
                    "type": "boolean"
                },
                "nickname": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
...
        "Message": {
            "type": "object",
            "properties": {
                "attachment": {
                    "$ref": "#/definitions/Attachment"
                },
                "createdAt": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "text": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                },
                "user": {
                    "$ref": "#/definitions/Member"
                }
            }
        },
        "MessageRequest": {
            "type": "object",
            "properties": {
                "file": {
                    "description": "image/* or audio/*",
                    "type": "string",
                    "format": "binary"
                },
                "text": {
                    "description": "Maximum 2000 characters",
                    "type": "string"
                }
            }
        },
...
        "SuccessResponse": {
            "type": "object",
            "properties": {
                "success": {
                    "description": "Only returns true, not a json object",
                    "type": "boolean"
                }
            }
        },
...
    }
}